Organize merchant endpoint tests in separate files
